+++ driver-pulsar/src/main/java/io/sbk/Pulsar/PulsarWriter.java
@@ -20,11 +20,10 @@ import org.apache.pulsar.client.api.PulsarClientException;
 /**
  * Class for Pulsar writer/producer.
  */
-public class PulsarWriter extends Writer {
+public class PulsarWriter implements Writer {
     final private Producer<byte[]> producer;
 
     public PulsarWriter(int writerID, Parameters params, String topicName, PulsarClient client) throws IOException {
-        super(writerID, params);
         try {
             this.producer = client.newProducer()
                     .enableBatching(true)
