+++ sbk-api/src/main/java/io/sbk/main/SbkMain.java
@@ -14,8 +14,9 @@ import io.sbk.api.Benchmark;
 import io.sbk.api.Parameters;
 import io.sbk.api.Performance;
 import io.sbk.api.SbkPerformance;
-import io.sbk.api.Writer;
-import io.sbk.api.Reader;
+import io.sbk.api.SbkReader;
+import io.sbk.api.SbkWriter;
+import io.sbk.api.QuadConsumer;
 import io.sbk.api.ResultLogger;
 import io.sbk.api.SystemResultLogger;
 
@@ -54,6 +55,8 @@ public class SbkMain {
         final ExecutorService executor;
         final Performance writeStats;
         final Performance readStats;
+        final QuadConsumer writeTime;
+        final QuadConsumer readTime;
 
         options.addOption("class", true, "Benchmark class");
         try {
@@ -119,10 +122,10 @@ public class SbkMain {
         if (params.writersCount > 0 && !params.writeAndRead) {
             writeStats = new SbkPerformance("Writing", REPORTINGINTERVAL, params.recordSize,
                                 params.csvFile, logger, executor);
-            params.recordWrite = writeStats::recordTime;
+            writeTime = writeStats::recordTime;
         } else {
             writeStats = null;
-            params.recordWrite = null;
+            writeTime = null;
         }
 
         if (params.readersCount > 0) {
@@ -134,21 +137,21 @@ public class SbkMain {
             }
             readStats = new SbkPerformance(action, REPORTINGINTERVAL, params.recordSize,
                             params.csvFile, logger, executor);
-            params.recordRead = readStats::recordTime;
+            readTime = readStats::recordTime;
         } else {
             readStats = null;
-            params.recordRead = null;
+            readTime = null;
         }
 
         try {
-            final List<Writer> writers =  IntStream.range(0, params.writersCount)
+            final List<SbkWriter> writers =  IntStream.range(0, params.writersCount)
                                             .boxed()
-                                            .map(i -> benchmark.createWriter(i, params))
+                                            .map(i -> new SbkWriter(i, params, writeTime, benchmark.createWriter(i, params)))
                                             .collect(Collectors.toList());
 
-            final List<Reader> readers = IntStream.range(0, params.readersCount)
+            final List<SbkReader> readers = IntStream.range(0, params.readersCount)
                                             .boxed()
-                                            .map(i -> benchmark.createReader(i, params))
+                                            .map(i -> new SbkReader(i, params, readTime, benchmark.createReader(i, params)))
                                             .collect(Collectors.toList());
 
             final List<Callable<Void>> workers = Stream.of(readers, writers)
